+++ api.php
@@ -40,7 +40,7 @@ if($_REQUEST['key'] && $_REQUEST['do']){
 	
 	if($_REQUEST['do'] == 'formulas'){
 
-		$sql = mysqli_query($conn, "SELECT name, notes, finalType AS concentration, fid, status, created, isProtected FROM formulasMetaData");
+		$sql = mysqli_query($conn, "SELECT name, notes, finalType AS concentration, fid, status, created, isProtected, rating FROM formulasMetaData");
 		while($r = mysqli_fetch_assoc($sql)) {
     		if (is_null($r['name']) || empty($r['name'])) {
         		$r['name'] = "N/A";
@@ -57,6 +57,7 @@ if($_REQUEST['key'] && $_REQUEST['do']){
 			$r['status'] = (int)$r['status'] ?: 0;
 			$r['created'] = (string)$C;
 			$r['isProtected'] = (int)$r['isProtected'] ?: 0;
+			$r['rating'] = (int)$r['rating'] ?: 0;
 
 			$rows[$_REQUEST['do']][] = array_filter($r);
 		}