+++ login.php
@@ -16,7 +16,6 @@ if(isset($_SESSION['parfumvault'])){
 require_once(__ROOT__.'/inc/config.php');
 require_once(__ROOT__.'/inc/opendb.php');
 require_once(__ROOT__.'/inc/product.php');
-require_once(__ROOT__.'/db/prepare/prepare.php');
 
 $installed_ver = mysqli_fetch_array(mysqli_query($conn,"SELECT app_ver FROM pv_meta"));
 
@@ -142,23 +141,6 @@ if($_POST['email'] && $_POST['password']){
  </body>
 </html>
 
-<!--PRE UPGRADE NEEDED-->
-<div class="modal fade" id="pre_warn" tabindex="-1" role="dialog" aria-labelledby="pre_warn" aria-hidden="true">
-  <div class="modal-dialog" role="document">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h5 class="modal-title">WARNING</h5>
-      </div>
-      <div id="upgrademsg"></div>
-      <div class="modal-body">
-		<?php echo prepare($installed_ver['app_ver'], $ver); ?>
-      </div>
-      <div class="modal-footer">
-        <button type="button" class="btn btn-primary" id="confUpgrade" data-dismiss="modal">Upgrade</button>
-      </div>
-    </div>
-  </div>
-</div>
 
 <!--FORGOT PASS INFO-->
 <div class="modal fade" id="forgot_pass" tabindex="-1" role="dialog" aria-labelledby="forgot_pass" aria-hidden="true">
@@ -196,35 +178,6 @@ $(document).ready(function() {
 		$('[data-toggle="tooltip"]').tooltip()
 	});
 
-<?php if($installed_ver['app_ver'] < $ver && !$first_time){ ?> 
-    $('#pre_warn').modal('show');
-	$("#pre_warn").on('hide.bs.modal', function () {  
-  		return false
-	});
-	$('#pre_warn').on('click', '[id*=confUpgrade]', function () {
-		
-		$.ajax({ 
-			url: '/db/prepare/prepare.php', 
-			type: 'POST',
-			data: {
-				action: 'upgrade',
-				version: <?=$ver?>,
-			},
-			dataType: 'json',
-			success: function (data) {
-				if (data.success){ 
-				    window.location='/'
-				}
-				if(data.error){
-					var msg = '<div class="alert alert-danger">'+data.error+'</div>';
-				}
-				
-				$('#upgrademsg').html(msg);
-			}
-		});
-																 
-	});
-<?php } ?>
 	$('#reg_form').on('click', '[id*=registerSubmit]', function () {
 		$('#registerSubmit').prop('disabled', true);
 		$('#msg').html('<div class="alert alert-info"><img src="/img/loading.gif"/> Please wait, configuring the system...<p><strong>Please do not close, refresh or navigate away from this page. You will be automatically redirected upon a succesfull installation.</strong></p></div>');